requirements.txt
disposable-email-domains==0.0.95
email-validator==2.1.0.post1
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0.post1